refactor(tools): phase 6 — cleanup, CHANGELOG, docs (final: 97 → 13)
Completes the tool consolidation series. Final surface: 13 tools (9
domain routers + 4 standalones), down from 97 — an 87% reduction.

Cleanup:
- Deleted no-op stubs: register_pcb_drc_fix_tools, register_netlist_tools
- Fixed stale internal docstrings referencing non-existent
  update_pcb_from_schematic (planned but never shipped; build_pcb_from_schematic
  covers the pipeline use case)
- Renamed schematic_router.py → schematic.py; old schematic.py → schematic_impl.py
  so router module names follow the same convention across all 9 routers

Docs:
- CHANGELOG.md: new file with full 84-entry rename mapping grouped by
  phase; called out as a breaking change (no compat aliases)
- README.md / AGENT-INSTALL.md / TOOLS.md: rewrote workflow examples
  to use router-style calls (router(operation="...", ...))

All 609 tests pass.
